server/requirements.txt
awscli==1.32.113
boto3==1.34.113
botocore==1.34.113
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
docker==3.4.0
docker-pycreds==0.4.0
docutils==0.16
filelock==3.14.0
fonttools==4.52.1
fsspec==2024.5.0
grpcio==1.64.0
grpcio-tools==1.64.0
idna==3.7
jaraco.classes==3.4.0
jaraco.context==5.3.0
Jinja2==3.1.4
jmespath==1.0.1
keyring==8.7
keyrings.alt==5.0.1
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.0
more-itertools==10.2.0
mpmath==1.3.0
networkx==3.3
opencv-python==4.9.0.80
numpy~=1.26.4
packaging==24.0
pafy==0.5.5
pandas==2.2.2
pillow==10.3.0
protobuf==5.26.1
psutil==5.9.8
py-cpuinfo==9.0.0
pyasn1==0.6.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.2
rsa==4.7.2
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
s3transfer==0.10.1
scipy==1.13.1
seaborn==0.13.2
setuptools==69.5.1
six==1.16.0
sympy==1.12
tabulate==0.9.0
thop==0.1.1.post2209072238
torch==2.3.0
torchvision==0.18.0
tqdm==4.66.4
typing_extensions==4.12.0
tzdata==2024.1
ultralytics==8.2.22
urllib3==2.2.1
voluptuous==0.14.2
websocket-client==1.8.0
wheel==0.43.0
yolo==0.3.2
youtube-dl==2020.12.2
